refactor: configure logic

+++ internal/testutil/mongolocal/mongolocal.go
@@ -4,6 +4,7 @@
 package mongolocal
 
 import (
+	"bufio"
 	"fmt"
 	"math/rand"
 	"net"
@@ -11,6 +12,10 @@ import (
 	"os/exec"
 	"path/filepath"
 	"testing"
+
+	testenv "github.com/01Joseph-Hwang10/terraform-provider-mongodb/internal/testutil/env"
+	"go.uber.org/zap"
+	"go.uber.org/zap/zaptest"
 )
 
 type Config struct {
@@ -30,59 +35,91 @@ func RandomPort() int {
 }
 
 func DefaultConfig() Config {
-	cwd, err := os.Getwd()
-	if err != nil {
-		panic(err)
-	}
+	// Get a random port.
 	port := RandomPort()
+
+	// Get DB path.
+	dbPath := filepath.Join(testenv.ExecRoot(), "tmp", "mongodb", fmt.Sprintf("db-%d", port))
+
 	return Config{
-		DbPath: filepath.Join(cwd, "tmp", "mongodb", fmt.Sprintf("db-%d", port)),
-		Port:   RandomPort(),
+		DbPath: dbPath,
+		Port:   port,
 	}
 }
 
 type MongoLocal struct {
-	proc   *exec.Cmd
-	config Config
+	proc    *exec.Cmd
+	watcher *exec.Cmd
+	config  Config
+	logger  *zap.Logger
 }
 
-func New(config Config) (*MongoLocal, error) {
-	// Ensure the path exists.
-	if err := os.MkdirAll(config.DbPath, 0755); err != nil {
-		return nil, err
-	}
-
+func New(t *testing.T, config Config) (*MongoLocal, error) {
+	// Create a command to start MongoDB.
 	cmd := exec.Command("mongod", "--dbpath", config.DbPath, "--port", fmt.Sprintf("%d", config.Port))
+
+	// Create a logger.
+	level := zap.InfoLevel
+	if testenv.IsDebug() {
+		level = zap.DebugLevel
+	}
+	logger := zaptest.NewLogger(t, zaptest.Level(level))
+
 	return &MongoLocal{
 		proc:   cmd,
 		config: config,
+		logger: logger,
 	}, nil
 }
 
-func WithMongoLocal(t *testing.T, callback func(*MongoLocal)) {
-	server, err := New(DefaultConfig())
-	t.Log("Starting MongoDB server...")
+func RunWithServer(t *testing.T, callback func(*MongoLocal)) {
+	// Create a new MongoDB server.
+	server, err := New(t, DefaultConfig())
+	logger := server.logger
+	logger.Info("Starting MongoDB server...")
 	if err != nil {
-		t.Fatal(err)
+		logger.Error("Failed to create MongoDB server", zap.Error(err))
 		return
 	}
 
+	// Attach logger to the server.
+	reader, err := server.Process().StdoutPipe()
+	if err != nil {
+		logger.Error("Failed to attach logger to MongoDB server", zap.Error(err))
+		return
+	}
+	go func() {
+		scanner := bufio.NewScanner(reader)
+		for scanner.Scan() {
+			logger.Sugar().Debugf("[mongod] %s", scanner.Text())
+		}
+	}()
+
+	// Start the server.
 	if err := server.Start(); err != nil {
-		t.Fatal(err)
+		logger.Error("Failed to start MongoDB server", zap.Error(err))
+		if err := server.Stop(); err != nil {
+			logger.Error("Failed to stop MongoDB server", zap.Error(err))
+		}
 		return
 	}
-	t.Logf("MongoDB server started on port %d", server.Config().Port)
+	logger.Sugar().Infof("MongoDB server started on port %d", server.Config().Port)
 
+	defer func() {
+		logger.Info("Stopping MongoDB server...")
+		if err := server.Stop(); err != nil {
+			logger.Error("Failed to stop MongoDB server", zap.Error(err))
+			return
+		}
+	}()
 	callback(server)
-
-	t.Log("Stopping MongoDB server...")
-	if err := server.Stop(); err != nil {
-		t.Fatal(err)
-		return
-	}
 }
 
-func (m *MongoLocal) clean() error {
+func (m *MongoLocal) createStorage() error {
+	return os.MkdirAll(m.config.DbPath, 0755)
+}
+
+func (m *MongoLocal) clearStorage() error {
 	return os.RemoveAll(m.config.DbPath)
 }
 
@@ -98,19 +135,47 @@ func (m *MongoLocal) URI() string {
 	return fmt.Sprintf("mongodb://localhost:%d", m.config.Port)
 }
 
+func (m *MongoLocal) Logger() *zap.Logger {
+	return m.logger
+}
+
 func (m *MongoLocal) Start() error {
-	if err := m.clean(); err != nil {
+	if err := m.clearStorage(); err != nil {
 		return err
 	}
-	return m.proc.Start()
+	if err := m.createStorage(); err != nil {
+		return err
+	}
+	if err := m.proc.Start(); err != nil {
+		return err
+	}
+
+	// Create a watcher to kill the process if the parent process dies.
+	m.watcher = exec.Command("/bin/sh", "-c", watcherScript(os.Getpid(), m.proc.Process.Pid))
+
+	if err := m.watcher.Start(); err != nil {
+		return err
+	}
+
+	return nil
 }
 
 func (m *MongoLocal) Stop() error {
-	if err := m.proc.Process.Kill(); err != nil {
-		return err
+	if m.proc.Process != nil {
+		if err := m.proc.Process.Kill(); err != nil {
+			m.logger.Warn("Failed to kill MongoDB process", zap.Error(err))
+		}
+		if err := m.proc.Wait(); err != nil {
+			m.logger.Warn("Failed to wait for MongoDB process", zap.Error(err))
+		}
 	}
-	if err := m.proc.Wait(); err != nil {
-		return err
+	if m.watcher.Process != nil {
+		if err := m.watcher.Process.Kill(); err != nil {
+			m.logger.Warn("Failed to kill MongoDB watcher process", zap.Error(err))
+		}
+		if err := m.watcher.Wait(); err != nil {
+			m.logger.Warn("Failed to wait for MongoDB watcher process", zap.Error(err))
+		}
 	}
-	return m.clean()
+	return m.clearStorage()
 }
